changes the structure by adding re-export
