+++ controllers/index.js
@@ -1,0 +1,15 @@
+const getAllContacts = require("./getAllContacts");
+const getById = require("./getById");
+const addContact = require("./addContact");
+const deleteContact = require("./deleteContact");
+const changeContact = require("./changeContact");
+const updateStatusContact = require("./updateStatusContact");
+
+module.exports = {
+	getAllContacts,
+	getById,
+	addContact,
+	deleteContact,
+	changeContact,
+	updateStatusContact,
+};
