+++ routes/api/contacts.js
@@ -2,7 +2,7 @@ const express = require("express");
 
 const router = express.Router();
 
-const controllers = require("../../controllers/contacts");
+const controllers = require("../../controllers");
 
 router.get("/", controllers.getAllContacts);
