+++ schemas/index.js
@@ -1,0 +1,3 @@
+const addSchema = require("./joi");
+
+module.exports = addSchema;
